rename(Board, BoardFactory): board 관련 패키지 생성 및 이동
