+++ src/main/java/chess/domain/piece/blank/Blank.java
@@ -1,6 +1,6 @@
 package chess.domain.piece.blank;
 
-import chess.domain.Board;
+import chess.domain.board.Board;
 import chess.domain.color.Color;
 import chess.domain.piece.Piece;
 import chess.domain.piece.PieceType;
