+++ src/main/java/chess/domain/piece/nonsliding/NonSlidingPiece.java
@@ -1,6 +1,6 @@
 package chess.domain.piece.nonsliding;
 
-import chess.domain.Board;
+import chess.domain.board.Board;
 import chess.domain.color.Color;
 import chess.domain.piece.Direction;
 import chess.domain.piece.Piece;
