+++ src/main/java/chess/domain/score/ScoreManager.java
@@ -1,6 +1,6 @@
 package chess.domain.score;
 
-import chess.domain.Board;
+import chess.domain.board.Board;
 import chess.domain.color.Color;
 import chess.domain.piece.Piece;
 import chess.domain.piece.PieceType;
