+++ src/main/java/chess/domain/state/BlankChessState.java
@@ -1,6 +1,6 @@
 package chess.domain.state;
 
-import chess.domain.Board;
+import chess.domain.board.Board;
 import chess.domain.color.Color;
 import chess.domain.position.Positions;
