+++ src/main/java/chess/domain/state/GeneralChessState.java
@@ -1,6 +1,6 @@
 package chess.domain.state;
 
-import chess.domain.Board;
+import chess.domain.board.Board;
 import chess.domain.color.Color;
 import chess.domain.piece.Piece;
 import chess.domain.position.Position;
