+++ src/test/java/chess/domain/board/BlankBoard.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.board;
 
 import chess.domain.piece.Piece;
 import chess.domain.position.Position;
