+++ src/test/java/chess/domain/state/GeneralChessStateTest.java
@@ -4,8 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 import static org.junit.jupiter.api.DynamicTest.dynamicTest;
 
-import chess.domain.BlankBoard;
-import chess.domain.Board;
+import chess.domain.board.BlankBoard;
+import chess.domain.board.Board;
 import chess.domain.color.Color;
 import chess.domain.piece.PieceType;
 import chess.domain.piece.nonsliding.King;
